Extract Tuple class and drop HasFunctionAddress
