+++ src/main/java/org/metricshub/jawk/intermediate/VersionManager.java
@@ -22,41 +22,35 @@ package org.metricshub.jawk.intermediate;
  * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
  */
 
-/**
- * An interface to a tuple position for interpretation.
- * <p>
- * This is differentiated from a position interface for
- * compilation because compilation requires linear
- * access (i.e., non-jumps) to the tuple list, while
- * interpretation requires this as well as jump capability.
- *
- * @author Danny Daglas
- */
-public interface PositionForInterpretation extends Position {
-	/**
-	 * Reposition to the tuple located at a particular address.
-	 * This is usually done in a response to an if condition.
-	 * However, this is also done to perform loops, etc.
-	 *
-	 * @param address The target address for the jump.
-	 */
-	void jump(Address address);
+import java.io.IOException;
+import java.io.InvalidClassException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.io.Serializable;
 
-	/**
-	 * <p>
-	 * current.
-	 * </p>
-	 *
-	 * @return The current index into the tuple list (queue)
-	 *         of the tuple located at the current position.
-	 */
-	int current();
+class VersionManager implements Serializable {
 
-	/**
-	 * Reposition to the tuple located at a particular index
-	 * into the tuple list (queue)..
-	 *
-	 * @param idx The target index for the jump.
-	 */
-	void jump(int idx);
+	private static final long serialVersionUID = -2015316238483923915L;
+
+	private static final int CLASS_VERSION = 2;
+
+	private int instanceVersion = CLASS_VERSION;
+
+	private void readObject(ObjectInputStream ois) throws IOException, ClassNotFoundException {
+		instanceVersion = ois.readInt();
+		if (instanceVersion != CLASS_VERSION) {
+			throw new InvalidClassException(
+					"Invalid intermeidate file format (instance version " + instanceVersion
+							+ " != class version " + CLASS_VERSION + ")");
+		}
+	}
+
+	private void writeObject(ObjectOutputStream oos) throws IOException {
+		oos.writeInt(instanceVersion);
+	}
+
+	@Override
+	public String toString() {
+		return "intermediate file format version = " + instanceVersion;
+	}
 }
